refactoring models and entity
