+++ Backend/app/main.py
@@ -1,8 +1,9 @@
 from fastapi import FastAPI, Depends, HTTPException,Request
 from sqlalchemy.orm import Session
 import logging
-from . import models, schemas, secrets
-from .services import  articleservice,service
+from . import  secrets
+from  app.dto import  schemas
+from .services import articleservice, service
 from .database import engine, Base, get_db
 from pydantic import BaseModel
 from fastapi import FastAPI, BackgroundTasks
